Refactored LLBase into engine directory
