+++ engine/objects/Agent.lua
@@ -1,15 +1,15 @@
 ------------------------------------------------------
--- Abstracts an inteligent agent for AI 
+-- Abstracts an inteligent agent for AI
 --
 -- @author jose@zunware.com
 -- @copyright Zunware
 ------------------------------------------------------
 local class 		= require "middleclass"							-- OOP
-local MovingObject  = require("LLBase.Renderer.MovingObject")		-- Base moving object
-local AIBehaviors   = require("LLBase.AI.Behaviors")				-- State Machines
+local MovingObject  = require("engine.objects.abstract.MovingObject")		-- Base moving object
+local AIBehaviors   = require("engine.ai.BehaviorFunctions")				-- State Machines
 
 -- The AI Class
-local IntelligentObject = class('IntelligentObject', MovingObject) 	--The Object inherits from a Moving Object 
+local IntelligentObject = class('IntelligentObject', MovingObject) 	--The Object inherits from a Moving Object
 
 -- Some flags
 local renderDebug   = false
@@ -38,7 +38,7 @@ function IntelligentObject:changeState(state)
 	-- Perform a change of state
 	local oldBehavior = self.currentBehavior.onExit(self) -- Exit last behavior
 	state.onEnter(self)									  -- Enter new behavior
-	self.currentBehavior = state 
+	self.currentBehavior = state
 end
 
 ------------------------------------------------------
@@ -50,7 +50,7 @@ function IntelligentObject:chill()
 end
 
 ------------------------------------------------------
--- Change to seek behavior. Follow the target 
+-- Change to seek behavior. Follow the target
 --
 -- @param target **(MovingObject)** The target
 ------------------------------------------------------
@@ -70,22 +70,22 @@ function IntelligentObject:lookAt(coord)
 	-- Get the targets position.
 	local targetX = coord.x
 	local targetY = coord.y
-	
+
 	-- Get a vector from us to the target (distVector)
 	local distX = targetX - self.x
 	local distY = targetY - self.y
-	
+
 	-- Figure out our forward vector (has to be normalized)
 	local fwd = {x = distX,y = distY}
-	
+
 	-- Normalize to set the heading vector.
 	local length = math.sqrt((distX ^ 2) + (distY ^ 2))
-	
+
 	if(length > 0) then
 		fwd.x = distX / length
 		fwd.y = distY / length
 	end
-	
+
 	-- Sets the heading to our forward vector.
 	self:setHeading(fwd.x, fwd.y)
 end
@@ -101,7 +101,7 @@ function IntelligentObject:update(dt)
 end
 
 ------------------------------------------------------
--- Returns the distance squared to given target 
+-- Returns the distance squared to given target
 --
 -- @return **(number)** The distance squared to target
 ------------------------------------------------------
